moved residue integrity methods from RNAResidue to moderna.analyze
