+++ tests/test_rna_residue.py
@@ -101,61 +101,6 @@ class RNAResidueTests(TestCase):
         self.assertTrue(resi1['N*'])
         resi2 = RNAResidue(chain[(' ', 16, ' ')])
         self.assertRaises(RNAResidueError, resi2.__getitem__, 'N*')
-        
-    # -------- TESTS FOR INTEGRITY ----------------
-    def test_is_backbone_complete(self):
-        """Missing backbone atoms are recognized."""
-        for resi in self.chain:
-            resi = RNAResidue(resi)
-            #self.assertTrue(resi.is_backbone_complete())
-        # .. and now the negative example.
-        chain=PDBParser().get_structure('test_struc',INCOMPLETE_BACKBONE)[0].child_list[0]
-        for resi in chain:
-            resi = RNAResidue(resi)
-            self.assertFalse(resi.is_backbone_complete())
-        
-    def test_is_backbone_intact(self):
-        """Should check all kinds of backbone discontinuities in one residue."""
-        chain=PDBParser().get_structure('test_struc',BROKEN_BACKBONE)[0].child_list[0]
-        residues = [r for r in chain]
-        for resi in residues[:5]:
-            mr = RNAResidue(resi)
-            self.assertFalse(mr.is_backbone_intact())
-        mr = RNAResidue(chain[6])
-        self.assertTrue(mr.is_backbone_intact())
-        
-    def test_is_backbone_intact_5p3p(self):
-        """Should check all kinds of backbone discontinuities in one residue."""
-        chain=PDBParser().get_structure('test_struc',BROKEN_BACKBONE)[0].child_list[0]
-        residues = [r for r in chain]
-        result_5p = []
-        result_3p = []
-        for resi in residues[:6]:
-            mr = RNAResidue(resi)
-            result_5p.append(mr.is_backbone_intact(mode="5'"))
-            result_3p.append(mr.is_backbone_intact(mode="3'"))
-        self.assertEqual(result_5p, [False, False, False, False, True, True])
-        self.assertEqual(result_3p, [True, True, True, False, False, True])
-        
-    def test_is_phosphate_intact(self):
-        """Should check whether OP1 and OP2 are in place"""
-        chain=PDBParser().get_structure('test_struc','test_data/rna_structures/bb_messed_up.pdb')[0].child_list[0]
-        resi1 = RNAResidue(chain[('H_c  ', 32, ' ')])
-        resi2 = RNAResidue(chain[(' ', 33, ' ')])
-        self.assertTrue(resi1.is_phosphate_intact())
-        self.assertFalse(resi2.is_phosphate_intact())
-    
-    def test_is_backbone_congested(self):
-        """Should check whether backbone atoms clash into rest of the structure."""
-        resi = RNAResidue(self.chain.child_list[2])
-        self.assertFalse(resi.is_backbone_congested())
-        # now check a structure where the backbone clashes into O2'
-        chain=PDBParser().get_structure('test_struc','test_data/rna_structures/bb_messed_up.pdb')[0].child_list[0]
-        resi = RNAResidue(chain[('H_c  ', 32, ' ')])
-        self.assertTrue(resi.is_backbone_congested())
-        resi = RNAResidue(chain[(' ', 33, ' ')])
-        self.assertTrue(resi.is_backbone_congested())
-
 
     def test_purine(self):
         """RNAResidue recognizes purines."""
